UIEH-1054 Fix comments
