+++ src/components/internal-link/index.js
@@ -1,0 +1,1 @@
+export { default } from './internal-link';
