+++ src/components/token-display/index.js
@@ -1,0 +1,1 @@
+export { default } from './token-display';
