+++ src/components/token-display/token-display.test.js
